Ported some scheduling routines from Java code

+++ model/job.go
@@ -2,7 +2,7 @@ package model
 
 type Job struct {
 	id       int
-	duration int
+	duration int64
 	history  []string
 	status   int // may change
 }
